renaming files
need to figure out how to show or hide the notification onClick of last social button
