+++ src/components/Notification/styles.module.scss
@@ -14,8 +14,8 @@
   top: -10px;
   right: 0px;
 
-  -webkit-animation: notificationslide 0.5s forwards;
-  animation: notificationslide 0.5s forwards;
+  -webkit-animation: notificationslide 0.7s forwards;
+  animation: notificationslide 0.7s forwards;
 }
 
 @-webkit-keyframes notificationslide {
